Avoid warnings from fct.h.

+++ fct.h
@@ -2480,10 +2480,10 @@ fct_standard_logger__on_fct_end(fct_logger_i *logger_, fctkern_t const *nk)
     num_passed = fctkern__tst_cnt_passed(nk);
 
     printf(
-        "%s (%d/%d tests",
+        "%s (%lu/%lu tests",
         (is_success) ? "PASSED" : "FAILED",
-        num_passed,
-        num_tests
+        (long unsigned)num_passed,
+        (long unsigned)num_tests
     );
 
     elasped_time = fct_timer__duration(&(logger->timer));
